backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.5.0
python-multipart==0.0.6
email-validator==2.0.0 